Taylor Swift Album Songs

A data frame containing details of songs from Taylor Swift's albums.

Usage

taylor_album_songs

Format

A data frame with the following columns:

X

Row identifier

album_name

Name of the album the song belongs to

ep

Indicates whether the album is an EP (Extended Play)

album_release

Release date of the album

track_number

Track number of the song within the album

track_name

Title of the song

artist

Name of the artist

featuring

Featured artists on the track, if any

bonus_track

Indicates if the song is a bonus track

promotional_release

Indicates if the song was released for promotional purposes

single_release

Indicates if the song was released as a single

track_release

Release date of the individual track

danceability

Danceability score (0.0 to 1.0)

energy

Energy score (0.0 to 1.0)

key

Musical key the song is in (numeric representation)

loudness

Loudness of the track in decibels (dB)

mode

Mode of the track (0 = Minor, 1 = Major)

speechiness

Speechiness score (0.0 to 1.0)

acousticness

Acousticness score (0.0 to 1.0)

instrumentalness

Instrumentalness score (0.0 to 1.0)

liveness

Liveness score (0.0 to 1.0)

valence

Valence score (0.0 to 1.0)

tempo

Tempo of the track in beats per minute (BPM)

time_signature

Time signature of the track (e.g., 4 for 4/4)

duration_ms

Duration of the track in milliseconds

explicit

Indicates if the track contains explicit lyrics

key_name

Musical key name (e.g., C major)

mode_name

Mode name (e.g., Major or Minor)

key_mode

Combined key and mode information

Examples

data(taylor_album_songs)
head(taylor_album_songs)
#>   X   album_name    ep album_release track_number             track_name
#> 1 1 Taylor Swift FALSE    2006-10-24            1             Tim McGraw
#> 2 2 Taylor Swift FALSE    2006-10-24            2        Picture To Burn
#> 3 3 Taylor Swift FALSE    2006-10-24            3 Teardrops On My Guitar
#> 4 4 Taylor Swift FALSE    2006-10-24            4  A Place In This World
#> 5 5 Taylor Swift FALSE    2006-10-24            5            Cold As You
#> 6 6 Taylor Swift FALSE    2006-10-24            6            The Outside
#>         artist featuring bonus_track promotional_release single_release
#> 1 Taylor Swift      <NA>       FALSE                <NA>     2006-06-19
#> 2 Taylor Swift      <NA>       FALSE                <NA>     2008-02-03
#> 3 Taylor Swift      <NA>       FALSE                <NA>     2007-02-19
#> 4 Taylor Swift      <NA>       FALSE                <NA>           <NA>
#> 5 Taylor Swift      <NA>       FALSE                <NA>           <NA>
#> 6 Taylor Swift      <NA>       FALSE                <NA>           <NA>
#>   track_release danceability energy key loudness mode speechiness acousticness
#> 1    2006-06-19        0.580  0.491   0   -6.462    1      0.0251      0.57500
#> 2    2006-10-24        0.658  0.877   7   -2.098    1      0.0323      0.17300
#> 3    2006-10-24        0.621  0.417  10   -6.941    1      0.0231      0.28800
#> 4    2006-10-24        0.576  0.777   9   -2.881    1      0.0324      0.05100
#> 5    2006-10-24        0.418  0.482   5   -5.769    1      0.0266      0.21700
#> 6    2006-10-24        0.589  0.805   5   -4.055    1      0.0293      0.00491
#>   instrumentalness liveness valence   tempo time_signature duration_ms explicit
#> 1                0   0.1210   0.425  76.009              4      232107    FALSE
#> 2                0   0.0962   0.821 105.586              4      173067    FALSE
#> 3                0   0.1190   0.289  99.953              4      203040    FALSE
#> 4                0   0.3200   0.428 115.028              4      199200    FALSE
#> 5                0   0.1230   0.261 175.558              4      239013    FALSE
#> 6                0   0.2400   0.591 112.982              4      207107    FALSE
#>   key_name mode_name key_mode
#> 1        C     major  C major
#> 2        G     major  G major
#> 3       A#     major A# major
#> 4        A     major  A major
#> 5        F     major  F major
#> 6        F     major  F major
summary(taylor_album_songs)
#>        X           album_name            ep          album_release     
#>  Min.   :  1.00   Length:194         Mode :logical   Length:194        
#>  1st Qu.: 49.25   Class :character   FALSE:194       Class :character  
#>  Median : 97.50   Mode  :character                   Mode  :character  
#>  Mean   : 97.50                                                        
#>  3rd Qu.:145.75                                                        
#>  Max.   :194.00                                                        
#>                                                                        
#>   track_number    track_name           artist           featuring        
#>  Min.   : 1.00   Length:194         Length:194         Length:194        
#>  1st Qu.: 5.00   Class :character   Class :character   Class :character  
#>  Median :10.00   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :10.71                                                           
#>  3rd Qu.:15.00                                                           
#>  Max.   :30.00                                                           
#>                                                                          
#>  bonus_track     promotional_release single_release     track_release     
#>  Mode :logical   Length:194          Length:194         Length:194        
#>  FALSE:171       Class :character    Class :character   Class :character  
#>  TRUE :23        Mode  :character    Mode  :character   Mode  :character  
#>                                                                           
#>                                                                           
#>                                                                           
#>                                                                           
#>   danceability       energy            key            loudness      
#>  Min.   :0.292   Min.   :0.1310   Min.   : 0.000   Min.   :-15.434  
#>  1st Qu.:0.511   1st Qu.:0.4465   1st Qu.: 2.000   1st Qu.: -9.326  
#>  Median :0.594   Median :0.5800   Median : 5.000   Median : -6.937  
#>  Mean   :0.584   Mean   :0.5745   Mean   : 4.686   Mean   : -7.518  
#>  3rd Qu.:0.652   3rd Qu.:0.7170   3rd Qu.: 7.000   3rd Qu.: -5.606  
#>  Max.   :0.897   Max.   :0.9500   Max.   :11.000   Max.   : -2.098  
#>  NA's   :3       NA's   :3        NA's   :3        NA's   :3        
#>       mode        speechiness       acousticness      instrumentalness   
#>  Min.   :0.000   Min.   :0.02310   Min.   :0.000191   Min.   :0.0000000  
#>  1st Qu.:1.000   1st Qu.:0.03080   1st Qu.:0.034600   1st Qu.:0.0000000  
#>  Median :1.000   Median :0.03960   Median :0.162000   Median :0.0000014  
#>  Mean   :0.911   Mean   :0.05831   Mean   :0.321225   Mean   :0.0039358  
#>  3rd Qu.:1.000   3rd Qu.:0.05740   3rd Qu.:0.662000   3rd Qu.:0.0000399  
#>  Max.   :1.000   Max.   :0.51900   Max.   :0.971000   Max.   :0.3480000  
#>  NA's   :3       NA's   :3         NA's   :3          NA's   :3          
#>     liveness          valence           tempo        time_signature 
#>  Min.   :0.03570   Min.   :0.0382   Min.   : 68.53   Min.   :1.000  
#>  1st Qu.:0.09295   1st Qu.:0.2535   1st Qu.: 99.98   1st Qu.:4.000  
#>  Median :0.11500   Median :0.4040   Median :121.96   Median :4.000  
#>  Mean   :0.14081   Mean   :0.4009   Mean   :125.99   Mean   :3.979  
#>  3rd Qu.:0.15050   3rd Qu.:0.5345   3rd Qu.:150.03   3rd Qu.:4.000  
#>  Max.   :0.59400   Max.   :0.9420   Max.   :208.92   Max.   :5.000  
#>  NA's   :3         NA's   :3        NA's   :3        NA's   :3      
#>   duration_ms      explicit         key_name          mode_name        
#>  Min.   :148781   Mode :logical   Length:194         Length:194        
#>  1st Qu.:209326   FALSE:172       Class :character   Class :character  
#>  Median :232107   TRUE :19        Mode  :character   Mode  :character  
#>  Mean   :237079   NA's :3                                              
#>  3rd Qu.:254448                                                        
#>  Max.   :613027                                                        
#>  NA's   :3                                                             
#>    key_mode        
#>  Length:194        
#>  Class :character  
#>  Mode  :character  
#>                    
#>                    
#>                    
#>